requirements.txt
customtkinter
pillow
tensorflow
numpy
opencv-python
